Split `sled_agent::sp` into two modules (general, sim)
